MM-3045: Create an assert for succesful searches
It includes a check if each parameter that was passed in the request URL
is handled by the server by inspecting the self link.
This is done in Groovy so that each parameter can automatically be
extracted from the search URL. It is also set to warning only, because
this mechanism might be too inprecise in practice. Still, it raises a
flag if the search parameter is absent from the self link.