MM-3266: Add TestScript for scenario 1.5, which is scenario 1.1 but with the care plan in free text format.
For internal executions, a "Nictiz-intern" target is built that is
basically the same as TestScript 1.1, because this scenario uses the
same content for the the same patient as with scenario 1.1, so all
queries will have the same result. For the production version, the only
difference is in the expected number of of Observation and Goal
resources.